Move document factory to own module
This feels like a bit more obvious locaiton that should facilitate
reuse.
