+++ network-api/networkapi/wagtailpages/factory/publication.py
@@ -3,7 +3,6 @@ from random import randint, shuffle, random
 
 from django.conf import settings
 
-from wagtail.core.models import Collection
 from wagtail_factories import PageFactory, ImageFactory
 from networkapi.wagtailpages.models import ArticlePage, Profile, PublicationPage
 from networkapi.utility.faker.helpers import get_homepage, reseed
@@ -11,18 +10,11 @@ from factory import (
     post_generation,
     Faker,
     SubFactory,
-    django,
 )
-from factory.django import DjangoModelFactory
 
 from networkapi.wagtailpages.pagemodels.publications.article import ArticleAuthors
+from networkapi.wagtailpages.factory.documents import DocumentFactory
 
-# UGLY COPYPASTE FROM latest
-# https://github.com/mvantellingen/wagtail-factories/blob/master/src/wagtail_factories/factories.py
-
-from wagtail_factories.factories import MP_NodeFactory
-
-from wagtail.documents import get_document_model
 
 RANDOM_SEED = settings.RANDOM_SEED
 TESTING = settings.TESTING
@@ -33,33 +25,6 @@ article_body_streamfield_fields = [
     'content',
     'full_width_image',
 ]
-
-
-class CollectionFactory(MP_NodeFactory):
-    name = Faker('text', max_nb_chars=60)
-
-    class Meta:
-        model = Collection
-
-
-class CollectionMemberFactory(DjangoModelFactory):
-    collection = SubFactory(CollectionFactory, parent=None)
-
-
-class DocumentFactory(CollectionMemberFactory):
-    class Meta:
-        model = get_document_model()
-
-    title = Faker('text', max_nb_chars=250)
-    file = django.FileField(
-        filename='test.pdf',
-        data=b'''
-              %PDF-1.0 1 0 obj<</Pages 2 0 R>>endobj
-              2 0 obj<</Kids[3 0 R]/Count 1>>endobj
-              3 0 obj<</MediaBox[0 0 3 3]>>endobj
-              trailer<</Root 1 0 R>>
-              '''
-    )
 
 
 class PublicationPageFactory(PageFactory):
